added confirm for deleting characters, delete button only visible when in edit mode

+++ src/4-CharacterSheet/DeleteButton.js
@@ -6,8 +6,10 @@ function DeleteButton({character}) {
   const navigate = useNavigate()
 
   const handleDelete = () => {
+    if ( window.confirm( "Are you sure?" ) ) {
       deleteLocalCharacter(character)
       navigate('/')
+    }
   }
 
   return (
